Should have command-handling code in command
This does not belong in IO.
