+++ walls-of-doom/command.h
@@ -1,6 +1,8 @@
 #ifndef COMMAND_H
 #define COMMAND_H
 
+#include "code.h"
+
 /**
  * The Command enumerated type represents the different commands the user may
  * issue.
@@ -18,5 +20,26 @@ typedef enum Command {
   COMMAND_QUIT,
   COMMAND_CLOSE
 } Command;
+
+/**
+ * Reads the next command that needs to be processed.
+ *
+ * This is the last pending command.
+ *
+ * This function consumes the whole input buffer and returns either
+ * COMMAND_NONE (if no other Command could be produced by what was in the input
+ * buffer) or the last Command different than COMMAND_NONE that could be
+ * produced by what was in the input buffer.
+ */
+Command read_next_command(void);
+
+/**
+ * Waits for the next command, blocking indefinitely.
+ */
+Command wait_for_next_command(void);
+/**
+ * Waits for any user input, blocking indefinitely.
+ */
+Code wait_for_input(void);
 
 #endif
